requirements.txt
annotated-types==0.6.0
anyio==4.2.0
async-timeout==4.0.3
asyncpg==0.30.0
click==8.1.7
exceptiongroup==1.2.0
fastapi==0.109.2
fastapi-utils==0.8.0
h11==0.14.0
idna==3.6
mypy_extensions==1.1.0
psutil==5.9.8
pyaes==1.6.1
pyasn1==0.5.1
pydantic==2.6.1
pydantic_core==2.16.2
python-dotenv==1.0.1
redis==5.0.1
rsa==4.9
sniffio==1.3.0
starlette==0.36.3
Telethon==1.34.0
typing-inspect==0.9.0
typing_extensions==4.9.0
uvicorn==0.27.1
